% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assign_node_attrs}
\alias{assign_node_attrs}
\title{Assign node attributes}
\usage{
assign_node_attrs(tree)
}
\arguments{
\item{tree}{Layer tree loaded from data_structure.yaml}
}
\value{
Nested list with assigned attributes for leafs and nodes
}
\description{
Assigns nodes in atlas_env$layer_options tree such that icons can be rendered 
correctly when layer selection tree is renderes in app sidebar
}
