% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_main_map_fct_BaseMap.R
\name{BaseMap}
\alias{BaseMap}
\title{Study site base map}
\usage{
BaseMap(shps = ImportSPA(db.connection = atlas_env$con))
}
\arguments{
\item{shps}{A named list of spatial files for the study area created with `ImportSPA`}
}
\value{
A leaflet map containing the outline and spatial elements of the ExeAtlas
 study site, including subcatchment outlines.
}
\description{
`BaseMap` returns a leaflet map that contains the outline and spatial elements
 of the ExeAtlas study area including site outlines and subcatchments
}
\details{
The function has no arguments but it loads data from the SQL database using
 the connection established by `CreateConnection()` when the app is first launched,
 or, ideally, relies on `SPA` shapefiles having been loaded into the parent environment already
 using the `ImportSPA()` function.
}
\examples{
\dontrun{
 p <- BaseMap()
}

}
