% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoastalHealth-package.R
\docType{package}
\name{CoastalHealth-package}
\alias{CoastalHealth}
\alias{CoastalHealth-package}
\title{CoastalHealth: Coastal Health Data Explorer}
\description{
A shiny-based intercative data discovery tool for marine and terrestrial data relating to coastal health in England. River catchment areas are based on EA operational catchment boundaries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mariolambrette/CoastalHealth}
  \item Report bugs at \url{https://github.com/mariolambrette/CoastalHealth/issues}
}

}
\author{
\strong{Maintainer}: Mario Lambrette \email{ml673@exeter.ac.uk}

}
\keyword{internal}
