% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_db_fct_db.R
\name{CreateConnection}
\alias{CreateConnection}
\title{Create SharePoint Database connection}
\usage{
CreateConnection(user)
}
\arguments{
\item{user}{UoE username to construct filepath}
}
\value{
An SQL connection to the temporary ExeAtlas database
}
\description{
Takes a UoE username as an input and connects to the temporary SQL database
on the ExeAtlas SharePoint. This is a temporary function until the database is
moved to a more permanent location.
}
\examples{
\dontrun{
  con <- CreateConnection('ml673')
}

}
\seealso{
[DBI::dbConnect()]
}
