% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_layerpopup_fct_createtable.R
\name{createtable}
\alias{createtable}
\title{Create reactable widget for rendering}
\usage{
createtable(layers, ns)
}
\arguments{
\item{layers}{Filtered verion of the layer_urls table containing user selected layers.}

\item{ns}{Module namepace for correct internal varibale namespacing}
}
\value{
a `reactable` widget that can be rendered by `reactable::renderReactable()`
}
\description{
Create reactable widget for rendering
}
\examples{
\dontrun{
layers <- read.csv(system.file("extdata", "layer_urls.csv", package = "CoastalHealth"))
createtable(layers[1:5,])
}

}
