% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_main_map_fct_PolyData.R
\name{CSD}
\alias{CSD}
\title{Get Cumulative sewage discharge summary message}
\usage{
CSD(data = tbl(atlas_env$con, "SHP_DAT_csd") \%>\% collect())
}
\arguments{
\item{data}{cumulative sewage discharge table from database}
}
\value{
Description of upstream cummulative sewage discharges
}
\description{
Get Cumulative sewage discharge summary message
}
