% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_main_map_fct_PolyData.R
\name{DominantLC}
\alias{DominantLC}
\title{Get dominant land cover for a selected polygon}
\usage{
DominantLC(poly_row = data)
}
\arguments{
\item{poly_row}{row from POLY_data for the corresponding polygon}
}
\value{
Nma eof the dominant land cover class in selected polygon
}
\description{
Get dominant land cover for a selected polygon
}
