% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExeAtlas-package.R
\docType{package}
\name{ExeAtlas-package}
\alias{ExeAtlas}
\alias{ExeAtlas-package}
\title{ExeAtlas: Interactive map for the ExeAtlas project}
\description{
A shiny-based interactive map covering the Exe and surrounding catchments. Provides access to a database of ecological, geological and polutant data for the catchments via a simple map interface.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mariolambrette/ExeAtlas}
  \item Report bugs at \url{https://github.com/mariolambrette/ExeAtlas/issues}
}

}
\author{
\strong{Maintainer}: Mario Lambrette \email{ml673@exeter.ac.uk}

}
\keyword{internal}
