% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_map_fct_GetLayers.R
\name{Get_marinearea}
\alias{Get_marinearea}
\title{Calculate marine area based on selected opcats}
\usage{
Get_marinearea()
}
\value{
NULL - stores an sf object describing the marine area to 
 atlas_env$marinearea
}
\description{
Calculate marine area based on selected opcats
}
