% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_map_fct_GetLayers.R
\name{Get_opcats}
\alias{Get_opcats}
\title{Get selected operational catchment outlines as sf polygons for adding to map}
\usage{
Get_opcats()
}
\value{
sf object with selected operational catchments represented as polygons
}
\description{
Get selected operational catchment outlines as sf polygons for adding to map
}
\examples{
\dontrun{
opcat.polygons <- Get_opcats()
}

}
