% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_map_fct_GetLayers.R
\name{Get_wbs}
\alias{Get_wbs}
\title{Get selected water body data as sf lines and polygons}
\usage{
Get_wbs()
}
\value{
NULL - atlas_env level reactive values containing river, lake and 
 boundary data are modified within the function to trigger plotting functions 
 in the map server function.
}
\description{
Get selected water body data as sf lines and polygons
}
