% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_main_map_fct_HighlightPolys.R
\name{HighlightPolys}
\alias{HighlightPolys}
\title{Create colour index layer for highlighted polygons}
\usage{
HighlightPolys(
  selectID = NULL,
  sc = ImportSPA(db.connection = atlas_env$con)$SHP_SPA_sc
)
}
\arguments{
\item{selectID}{The ID of the polygon selected by the user}

\item{sc}{The subcatchment sf object (retrieved from the shps object imported into parent environment for generating basemaps)}
}
\value{
Subcatchment sf object with colour column
}
\description{
Function to be run when a polygon is clicked by user. returns an sf object
that can be added to the map with a column 'colours' that will highlight
the correct polygons upstream and downstream.
}
\examples{

p <- HighlightPolys(selectID = 53)

}
