% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_main_map_fct_BaseMap.R
\name{ImportSPA}
\alias{ImportSPA}
\title{Import spatial shapefiles from database}
\usage{
ImportSPA(db.connection)
}
\arguments{
\item{db.connection}{SQLite Connection to ExeAtlas database}
}
\value{
Named list of sf objects relating to the `SPA` category data in the specified database
}
\description{
Imports all the 'spatial' (`SPA`) category shapefiles from the SQL database and returns them in a named list.
These are the basic files that make up the outline of the map (e.g. background layers and catchment outlines)
}
\examples{
\dontrun{
 shps <- ImportSPA(con)
}

}
