% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_main_map_fct_PolyData.R
\name{KS_list}
\alias{KS_list}
\title{Skeleton for function to return WKT key species present list. currently returns NA}
\usage{
KS_list(polyID, poly_row = data)
}
\arguments{
\item{polyID}{selected polygon ID}

\item{poly_row}{Row of polygon data}
}
\value{
NA
}
\description{
Skeleton for function to return WKT key species present list. currently returns NA
}
