% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_layerpopup_fct_sf_handling.R
\name{load_sf}
\alias{load_sf}
\title{Load sf object to global environment}
\usage{
load_sf(id, url)
}
\arguments{
\item{id}{String denoting the varibale name to assign the sf object to}

\item{url}{String denoting the url to provide to sf for loading}
}
\value{
NULL - assigns result to global environment
}
\description{
Load sf object to global environment
}
\examples{
\dontrun{
 load_sf("example_layer", "https://ows.emodnet-humanactivities.eu/wfs?request
   =GetFeature&service=WFS&version=1.1.0&outputFormat=application\%2Fjson&type
   Name=emodnet:shellfish")
}

}
