% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_main_map_fct_AddLayers.R
\name{LoadLayer}
\alias{LoadLayer}
\title{Load spatial layers from database}
\usage{
LoadLayer(name, db.connection = atlas_env$con)
}
\arguments{
\item{name}{Name of the database table containing the spatial data}

\item{db.connection}{RSQLite database connection object}
}
\value{
sf object containing data layer
}
\description{
Uses the table name, supplied to the parameter `name` and a database connection
to load spatial layers stored in the database, returning an sf object ready for
plotting. Relies on spatial data having been loaded into the database as an
sf object, projected in wgs84 (epsg: 4326) with the geometry converted to WKT
format in a column called 'geom'
}
