% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_main_map_fct_PolyData.R
\name{MAs_list}
\alias{MAs_list}
\title{Get WKT format list of management areas present}
\usage{
MAs_list(poly_row = data)
}
\arguments{
\item{poly_row}{row from POLY_data for the corresponding polygon}
}
\value{
List of Management areas present in selected subcatchment
}
\description{
Get WKT format list of management areas present
}
