% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_map_fct_PlotLayers.R
\name{Plot_marinearea}
\alias{Plot_marinearea}
\title{Plot marine area}
\usage{
Plot_marinearea(map_proxy)
}
\arguments{
\item{map_proxy}{leaflet proxy object relating to the main map on which the
layer will be plotted}
}
\value{
NULL - plots marine area directly onto leaflet map given by map_proxy
}
\description{
This function plots the marine area onto the existing leaflet basemap. It is 
seperate to the waterbody family functions as the data displayed is calculated 
rather than sourced from the EA but in is executed in the same way from the 
wbview module following the user ticking the marine area box.
}
