% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_map_fct_PlotLayers.R
\name{Plot_opcats}
\alias{Plot_opcats}
\title{Plot selected operational catchments}
\usage{
Plot_opcats(map_proxy)
}
\arguments{
\item{map_proxy}{leaflet proxy object relating to the main map on which the
layer will be plotted}
}
\value{
NULL - updates the existing leaflet map
}
\description{
Plot selected operational catchments
}
\examples{
\dontrun{
Plot_opcats()
}

}
