% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_map_fct_PlotLayers.R
\name{Plot_wbs_rivers}
\alias{Plot_wbs_rivers}
\title{Family of functions to plot waterbody sptial data}
\usage{
Plot_wbs_rivers(map_proxy)
}
\arguments{
\item{map_proxy}{leaflet proxy object relating to the main map on which the
layer will be plotted}
}
\value{
NULL - updates the existing map
}
\description{
The below functions display the various spatial ekements of waterbodies (rivers,
lakes and the actual water body boundaries). These functions are triggered
by user inputs to checkboxes in the main app sidepanel
}
