% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_layerpopup_fct_createtable.R
\name{process_url}
\alias{process_url}
\title{Process urls to add variables}
\usage{
process_url(url)
}
\arguments{
\item{url}{A web url where any variables follow the above stated format}
}
\value{
Process urls as character strings
}
\description{
Helper function to process web urls to add variables such as opcat ids and 
spatial and temporal extents

Variables should follow the below format:

**operational catchmnet id:** {opcat_id}
**Spatial extent:** {xmin}, {xmax} etc.
**Temporal extent:** {sd_YYYY-MM-DD}, {ed_YYYY-MM-DD}. Replace with the appropriate format to be used as a character string in R datetime processing functions
}
\examples{
\dontrun{
 process_url("https://environment.data.gov.uk/geoservices/datasets/6da82900
 -d465-11e4-8cc3-f0def148f590/ogc/features/v1/collections/Saltmarsh_Extents_
 and_Zonation/items?limit=300000&bbox={xmin},{ymin},{xmax},{xmin}")
}
}
