% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{st_erase}
\alias{st_erase}
\title{Erase overlapping feature geometries}
\usage{
st_erase(x, y)
}
\arguments{
\item{x}{an sf object}

\item{y}{an sf object}
}
\value{
sf object containing all geometries in x that do not overlap with y
}
\description{
An extension of sf::st_difference that erases all parts of x that overlap with y
}
