% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_main_map_fct_AddLayers.R
\name{AddLayers}
\alias{AddLayers}
\title{Add layers to base map}
\usage{
AddLayers(add_layers = "none", db.connection = atlas_env$con)
}
\arguments{
\item{add_layers}{A list of file paths to the layers to add}

\item{bm}{Basic basemap created with the BaseMap() function. Calls `BaseMap()` by default, though this not the ideal implementation when the function is used in production}
}
\value{
A character string that contains the plotting command to be evaluated by `eval(parse(text = cmd))` to add the correct data layer to the base map
}
\description{
Add layers onto the basic base map, including optional data layers
}
\examples{

\dontrun{
print(ggiraph::girafe(ggobj = AddLayers(bm)))

# Complete plot
p <- BaseMap()
p1 <- AddLayers(bm = p)
p2 <- ggiraph::girafe(ggobj = p1)
print(p2)
}

}
