% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_map_fct_BaseMap.R
\name{BaseMap}
\alias{BaseMap}
\title{Leaflet basemap for Coastal Health data explorer}
\usage{
BaseMap()
}
\value{
A leaflet base map to be modified using `leafletProxy`
}
\description{
`BaseMap` returns a leaflet map that contains a basemap of england (either as
OpenStreetMap, the ESRI grey canvas, or a plain outline of the UK mainland) 
with placeholders for other data layers to be displayed. The function has no 
arguments but the returned map can be modified by using `leafletProxy`
}
\examples{
map <- BaseMap()

}
