% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{recentre_map}
\alias{recentre_map}
\title{Recentre map view}
\usage{
recentre_map(map_proxy)
}
\value{
NULL - updates the view of an existing map
}
\description{
Function to recentre map view to the correct extent. Correct extent should be
stored as an st_bbox object at atlas_env$bounds
}
\examples{
\dontrun{
 output$map <- BaseMap()
 
 recentre_map(leaflet::leafletProxy("map", session), atlas_env$map_bounds)
}

}
